network_inspector/requirements.txt
NetfilterQueue
scapy
requests
pyzmq
